+++ packages/openjpeg/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.2](https://localhost/compare/@cornerstonejs/codec-openjpeg@1.3.1...@cornerstonejs/codec-openjpeg@1.3.2) (2026-08-13)
+
+**Note:** Version bump only for package @cornerstonejs/codec-openjpeg
+
+
+
+
+
 ## [1.3.1](https://localhost/compare/@cornerstonejs/codec-openjpeg@1.3.0...@cornerstonejs/codec-openjpeg@1.3.1) (2026-07-09)